+++ pkg/dir/createdir.go
@@ -20,11 +20,3 @@ func CreateFile(filename string) {
 		}
 	}
 }
-
-func removeDir(dirname string) {
-	if _, err := os.Stat(dirname); !os.IsNotExist(err) {
-		if err := os.RemoveAll(dirname); err != nil {
-			fmt.Println("Error: ", err)
-		}
-	}
-}
