+++ pkg/dir/initdir.go
@@ -1,30 +1,10 @@
 package dir
 
-import (
-	"fmt"
-	"os"
-)
-
 func InitializeDir() {
 	// Create the .sgit/objects directory
-
-	if err := os.MkdirAll(".sgit/objects", 0755); err != nil {
-		fmt.Println("Error: ", err)
-	}
+	CreateDir(".sgit/objects")
 
 	// Create the sub directories
-	if err := os.Mkdir(".sgit/objects/info", 0755); err != nil {
-		fmt.Println("Error: ", err)
-	}
-
-	if err := os.Mkdir(".sgit/objects/pack", 0755); err != nil {
-		fmt.Println("Error: ", err)
-	}
-}
-
-func RemoveAllDir() {
-	// Remove the .sgit directory
-	if err := os.RemoveAll(".sgit"); err != nil {
-		fmt.Println("Error: ", err)
-	}
+	CreateDir(".sgit/objects/info")
+	CreateDir(".sgit/objects/pack")
 }
